small fixes

+++ app/src/main/java/ru/is2si/sisi/presentation/points/PointsPresenter.kt
@@ -36,7 +36,7 @@ class PointsPresenter @Inject constructor(
 
     override fun addPoint(pointName: String) {
         view.showLoading()
-        disposables += saveSelectPoint.execute(SaveSelectPoint.Params(pointName.toInt()))
+        disposables += saveSelectPoint.execute(SaveSelectPoint.Params(pointName))
                 .map { it.map { point -> point.asView() } }
                 .subscribeOn(rxSchedulers.io)
                 .observeOn(rxSchedulers.ui)
